ReadRules: viewer check fix

+++ src/modules/EPenser.js
@@ -161,13 +161,7 @@ export default class EPenser {
 
 		if (channel.id !== readRules.channel) return;
 
-		const roles = member.roles.array();
-		if (
-			roles.length !== 1 &&
-			(roles.length !== 2 || roles[1].id !== readRules.viewerRole)
-		)
-			//account for @everyone role
-			return message.delete();
+		if (member.roles.get(readRules.viewerRole)) return message.delete();
 
 		const announce =
 			readRules.announce && guild.channels.get(readRules.announceChannel);
